UPDATE: fix the step save_data_in_snowflake_using_copy
Why this change was necessary?
We were using the .to_json() function from pandas, which for some dataframes
generate an inconsistent view of the data, which makes the data arrives at
snowflake in a single variant column because snowpark is unable to infer the
schema of the table.

How does it address the problem?
- By using to_parquet() instead of to_json.
- We had some conversion issues using the default engine (pyarrow), so,
I decided to used the 'fastparquet' engine instead.

Related tickets or issues: [SOLENG-5620]