UPDATE: updating the step configuration to an output configuration
Why this change was necessary?
To be more flexible about how the data transits in stepsfera.

Now, the old behavior is kept (traffic the data using orchest.output)
But, a new output_type is available (output to file).

In the name of stability, the output format will be in parquet format.