FEAT: create step for automatic flattening of object columns in snowflake
Why this change was necessary?
To make easier to clean nested data in snowflake and be more
efficient in our daily analyses.

How does it address the problem:
- Identify columns that are objects in snowflake.
- Identify the existing keys in each column
- Ignore cases where it's not possible to flatten the column for some reason.
- Columns provenient from a flatten process will be store as string initially. In the future
we may improve to have an way of detecting the data type as well.
- Create select expressions and execute in snowflake using snowpark.

Any other relevant information:
- Add some preprocessing functions because flattening may exceed the amount of columns supported
on snowflake (2000). When applying the function without removing the stage columns we have trespassed
this limit.
Related tickets or issues: [INOX-1803]