UPDATE: fix snowflake_flatten_column step
Why this change was necessary?
Some issues with object columns with keys splitted by "-", for example,
associated_company={"company-id":"a"}

There was an issue in the conversion where the select expression
becomes associated_company"company-id and it should
associated_company__company_id in the alias.

Any other relevant information:
create a step that retrieves a sql statement as data frame.